unify GOOS=tamago amd64/arm/riscv64 runtime support
